+++ src/clj/lambdacd/internal/default_pipeline_state.clj
@@ -3,9 +3,7 @@
   i.e. what's currently running, what are the results of each step, ..."
   (:require [lambdacd.internal.default-pipeline-state-persistence :as persistence]
             [clj-time.core :as t]
-            [clojure.core.async :as async]
-            [lambdacd.internal.pipeline-state :as pipeline-state-protocol]
-            [lambdacd.event-bus :as event-bus]
+            [lambdacd.internal.pipeline-state :as pipeline-state]
             [lambdacd.util :as util]))
 
 (def clean-pipeline-state {})
@@ -41,7 +39,7 @@
 
 
 (defrecord DefaultPipelineState [state-atom home-dir]
-  pipeline-state-protocol/PipelineStateComponent
+  pipeline-state/PipelineStateComponent
   (update [self build-number step-id step-result]
     (update-legacy build-number step-id step-result home-dir state-atom))
   (get-all [self]
@@ -51,19 +49,8 @@
   (next-build-number [self]
     (next-build-number-legacy state-atom)))
 
-(defn- start-pipeline-state-updater [instance ctx] ; TODO could be generalized for others to use
-  (let [subscription (event-bus/subscribe ctx :step-result-updated)
-        step-updates-channel (util/buffered (event-bus/only-payload subscription))]
-    (async/go-loop []
-      (if-let [step-result-update (async/<! step-updates-channel)]
-        (let [step-result (:step-result step-result-update)
-              build-number (:build-number step-result-update)
-              step-id (:step-id step-result-update)]
-          (pipeline-state-protocol/update instance build-number step-id step-result)
-          (recur))))))
-
 (defn new-default-pipeline-state [state-atom config ctx]
   (let [home-dir (:home-dir config)
         instance (->DefaultPipelineState state-atom home-dir)]
-    (start-pipeline-state-updater instance ctx)
+    (pipeline-state/start-pipeline-state-updater instance ctx)
     instance))