+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The official release will have a defined and more stable API. If you are already
 
 ## 0.4.3
 
+Housekeeping release: Contains mostly cleanup under the hood and changes to APIs for advanced users.
+ If you were using custom control-flow steps, runners, persistence mechanisms or other advanced features, make sure you
+ look through the changes and upgrade as future releases will remove deprecated functionality.
+
 * Improvements:
   * Now providing events `:step-result-updated` and `:step-finished` on event-bus for use by other components like
     runners and persistence components